Modified the appengine-web.xml to adjust the following configuration parameter:
<async-session-persistence enabled="true" />

to the following: 

<async-session-persistence enabled="false" />

Using the asynchronous session persistence features in some real-world applications caused a seemingly endless stream of Task queue records to be created eventually exhausting the free quota for the application.  Until I have the chance to track down the root cause of this problem, I'm recommending against its use.